Replace -ffast-math with more glanular options
-ffast-math implies -ffinite-math-only, which disables NaN detection.
It may cause undefined behavior if nan is converted to int, and cause
confusion when debugging - isnan will always return false and thus
all code depending on it will be removed. So, we replace -ffast-math
with "-funsafe-math-optimizations -fno-math-errno -fno-trapping-math"
which should provide comparable performance and allow using isnan.
For safety, cause compilation error if -ffast-math is detected.